Add files via upload
Got rid of the Hertz low vs high and figuring out the shape broadcasting issues. in the standardisation of the model, it was giving and array of [3,1], which were each 1D arrays of the standaradised data. changed this to standardise() from function file so its treated in the same way as the ydatas. Still need to figure out the 4,56 and 3,56 shape incompatibility, but its fine